Rename preload_syscall.c/h to shim_syscall.c/h
